+++ src/render/cdSimpleAlphaShader.h
@@ -74,8 +74,8 @@ namespace cdSimpleAlphaPixel_Fullbright {
 // ============================================================================
 // Externs
 // ============================================================================
-extern void nglDxRegisterVShader(unsigned long* VS, const unsigned int* Microcode);  // ngl_dx_shader.o
-extern void nglDxRegisterPShader(unsigned long** PS, const unsigned int* Microcode);  // ngl_dx_shader.o
+extern void nglDxRegisterVShader(unsigned int* VS, const unsigned int* Microcode);  // ngl_dx_shader.o
+extern void nglDxRegisterPShader(unsigned int** PS, const unsigned int* Microcode);  // ngl_dx_shader.o
 
 extern cdSimpleAlphaShader* gCDSimpleAlphaShader;  // @0x10DE058
